Move Product Variant stocks migrations to bulk_migrations
