+++ saleor/graphql/product/mutations/products.py
@@ -30,11 +30,10 @@ from ....product.utils.attributes import (
 )
 from ....warehouse.error_codes import StockErorrCode
 from ....warehouse.management import set_stock_quantity
-from ....warehouse.models import Stock
 from ...core.mutations import BaseMutation, ModelDeleteMutation, ModelMutation
 from ...core.scalars import Decimal, WeightScalar
 from ...core.types import SeoInput, Upload
-from ...core.types.common import ProductError, StockError
+from ...core.types.common import ProductError
 from ...core.utils import (
     clean_seo_fields,
     from_global_id_strict_type,
@@ -43,7 +42,6 @@ from ...core.utils import (
 )
 from ...core.utils.reordering import perform_reordering
 from ...meta.deprecated.mutations import ClearMetaBaseMutation, UpdateMetaBaseMutation
-from ...utils import resolve_global_ids_to_primary_keys
 from ...warehouse.types import Warehouse
 from ..types import (
     Category,
@@ -1337,119 +1335,6 @@ class ProductVariantClearPrivateMeta(ClearMetaBaseMutation):
         public = False
         error_type_class = ProductError
         error_type_field = "product_errors"
-
-
-class ProductVariantStocksCreate(BaseMutation):
-    product_variant = graphene.Field(
-        ProductVariant, description="Updated product variant."
-    )
-
-    class Arguments:
-        variant_id = graphene.ID(
-            required=True,
-            description="ID of a product variant for which stocks will be created.",
-        )
-        stocks = graphene.List(
-            graphene.NonNull(StockInput),
-            required=True,
-            description="Input list of stocks to create.",
-        )
-
-    class Meta:
-        description = "Creates stocks for product variant."
-        permissions = (ProductPermissions.MANAGE_PRODUCTS,)
-        error_type_class = StockError
-        error_type_field = "stock_errors"
-
-    @classmethod
-    def perform_mutation(cls, root, info, **data):
-        stocks = data["stocks"]
-        variant = cls.get_node_or_error(
-            info, data["variant_id"], only_type=ProductVariant
-        )
-        warehouse_ids = [stock["warehouse"] for stock in stocks]
-        warehouses = cls.get_nodes_or_error(
-            warehouse_ids, "warehouse", only_type=Warehouse
-        )
-        try:
-            create_stocks(variant, stocks, warehouses)
-        except ValidationError as error:
-            error.code = StockErorrCode.UNIQUE
-            raise ValidationError({"warehouse": error})
-        return cls(product_variant=variant)
-
-
-class ProductVariantStocksUpdate(BaseMutation):
-    product_variant = graphene.Field(
-        ProductVariant, description="Updated product variant."
-    )
-
-    class Arguments:
-        variant_id = graphene.ID(
-            required=True,
-            description="ID of product variant for which stocks will be updated.",
-        )
-        stocks = graphene.List(
-            graphene.NonNull(StockInput),
-            required=True,
-            description="Input list of stocks.",
-        )
-
-    class Meta:
-        description = "Update stocks for product variant."
-        permissions = (ProductPermissions.MANAGE_PRODUCTS,)
-        error_type_class = StockError
-        error_type_field = "stock_errors"
-
-    @classmethod
-    def perform_mutation(cls, root, info, **data):
-        stocks = data["stocks"]
-        variant = cls.get_node_or_error(
-            info, data["variant_id"], only_type=ProductVariant
-        )
-        warehouse_ids = [stock["warehouse"] for stock in stocks]
-        warehouses = cls.get_nodes_or_error(
-            warehouse_ids, "warehouse", only_type=Warehouse
-        )
-        for stock_data, warehouse in zip(stocks, warehouses):
-            stock, _ = Stock.objects.get_or_create(
-                product_variant=variant, warehouse=warehouse
-            )
-            stock.quantity = stock_data["quantity"]
-            stock.save(update_fields=["quantity"])
-        return cls(product_variant=variant)
-
-
-class ProductVariantStocksDelete(BaseMutation):
-    product_variant = graphene.Field(
-        ProductVariant, description="Updated product variant."
-    )
-
-    class Arguments:
-        variant_id = graphene.ID(
-            required=True,
-            description="ID of product variant for which stocks will be deleted.",
-        )
-        warehouse_ids = graphene.List(graphene.NonNull(graphene.ID),)
-
-    class Meta:
-        description = "Delete stocks from product variant."
-        permissions = (ProductPermissions.MANAGE_PRODUCTS,)
-        error_type_class = StockError
-        error_type_field = "stock_errors"
-
-    @classmethod
-    def perform_mutation(cls, root, info, **data):
-        variant = cls.get_node_or_error(
-            info, data["variant_id"], only_type=ProductVariant
-        )
-        _, warehouses_pks = resolve_global_ids_to_primary_keys(
-            data["warehouse_ids"], Warehouse
-        )
-        Stock.objects.filter(
-            product_variant=variant, warehouse__pk__in=warehouses_pks
-        ).delete()
-        return cls(product_variant=variant)
 
 
 class ProductTypeInput(graphene.InputObjectType):
