+++ saleor/graphql/product/schema.py
@@ -23,6 +23,9 @@ from .bulk_mutations.products import (
     ProductTypeBulkDelete,
     ProductVariantBulkCreate,
     ProductVariantBulkDelete,
+    ProductVariantStocksBulkCreate,
+    ProductVariantStocksBulkDelete,
+    ProductVariantStocksBulkUpdate,
 )
 from .enums import StockAvailability
 from .filters import (
@@ -94,9 +97,6 @@ from .mutations.products import (
     ProductVariantClearPrivateMeta,
     ProductVariantCreate,
     ProductVariantDelete,
-    ProductVariantStocksCreate,
-    ProductVariantStocksDelete,
-    ProductVariantStocksUpdate,
     ProductVariantUpdate,
     ProductVariantUpdateMeta,
     ProductVariantUpdatePrivateMeta,
@@ -464,9 +464,9 @@ class ProductMutations(graphene.ObjectType):
     product_variant_delete = ProductVariantDelete.Field()
     product_variant_bulk_create = ProductVariantBulkCreate.Field()
     product_variant_bulk_delete = ProductVariantBulkDelete.Field()
-    product_variant_stocks_create = ProductVariantStocksCreate.Field()
-    product_variant_stocks_delete = ProductVariantStocksDelete.Field()
-    product_variant_stocks_update = ProductVariantStocksUpdate.Field()
+    product_variant_stocks_bulk_create = ProductVariantStocksBulkCreate.Field()
+    product_variant_stocks_bulk_delete = ProductVariantStocksBulkDelete.Field()
+    product_variant_stocks_bulk_update = ProductVariantStocksBulkUpdate.Field()
     product_variant_update = ProductVariantUpdate.Field()
     product_variant_translate = ProductVariantTranslate.Field()
     product_variant_update_metadata = ProductVariantUpdateMeta.Field(
